@@ -2,8 +2,6 @@
   "name": "@blackbaud/skyux-lib-clipboard",
   "version": "2.0.0-rc.0",
   "description": "Single-page-application for skyux-lib-clipboard",
-  "main": "bundles/bundle.umd.js",
-  "module": "index.ts",
   "scripts": {
     "test": "skyux test --coverage library",
     "build": "skyux build-public-library",
@@ -13,49 +11,46 @@
   "author": "Blackbaud",
   "license": "ISC",
   "peerDependencies": {
-    "@angular/common": "^9.0.1",
-    "@angular/core": "^9.0.1",
-    "@skyux/core": "^4.0.0-rc.0",
-    "@skyux/i18n": "^4.0.0-rc.0"
+    "@angular/common": "^9.1.9",
+    "@angular/core": "^9.1.9",
+    "@skyux/core": "^4.1.0",
+    "@skyux/i18n": "^4.0.0"
   },
   "dependencies": {
     "clipboard-polyfill": "2.8.6"
   },
   "devDependencies": {
-    "@angular/animations": "9.0.1",
-    "@angular/common": "9.0.1",
-    "@angular/compiler": "9.0.1",
-    "@angular/compiler-cli": "9.0.1",
-    "@angular/core": "9.0.1",
-    "@angular/forms": "9.0.1",
-    "@angular/platform-browser": "9.0.1",
-    "@angular/platform-browser-dynamic": "9.0.1",
-    "@angular/router": "9.0.1",
-    "@blackbaud/auth-client": "2.22.0",
+    "@angular/animations": "9.1.9",
+    "@angular/common": "9.1.9",
+    "@angular/compiler": "9.1.9",
+    "@angular/compiler-cli": "9.1.9",
+    "@angular/core": "9.1.9",
+    "@angular/forms": "9.1.9",
+    "@angular/platform-browser": "9.1.9",
+    "@angular/platform-browser-dynamic": "9.1.9",
+    "@angular/router": "9.1.9",
+    "@blackbaud/auth-client": "2.25.0",
     "@blackbaud/skyux-builder-plugin-code-block": "1.1.0",
-    "@pact-foundation/pact": "9.7.0",
-    "@pact-foundation/pact-web": "9.7.0",
-    "@skyux-sdk/builder": "4.0.0-rc.2",
-    "@skyux-sdk/builder-plugin-skyux": "1.3.1",
-    "@skyux-sdk/e2e": "3.1.2",
-    "@skyux-sdk/pact": "4.0.0-rc.0",
-    "@skyux-sdk/testing": "3.1.0",
-    "@skyux/assets": "4.0.0-rc.0",
-    "@skyux/config": "4.0.0-rc.1",
-    "@skyux/core": "4.0.0-rc.1",
-    "@skyux/http": "4.0.0-rc.2",
-    "@skyux/i18n": "4.0.0-rc.1",
-    "@skyux/indicators": "4.0.0-rc.2",
-    "@skyux/modals": "4.0.0-rc.2",
-    "@skyux/omnibar-interop": "4.0.0-rc.0",
-    "@skyux/router": "4.0.0-rc.0",
-    "@skyux/theme": "4.0.0-rc.1",
-    "codelyzer": "5.2.1",
-    "rxjs": "6.5.4",
-    "ts-node": "8.6.2",
-    "tslint": "5.20.1",
-    "typescript": "3.6.4",
-    "zone.js": "0.10.2"
+    "@skyux-sdk/builder": "4.0.0-rc.11",
+    "@skyux-sdk/builder-plugin-skyux": "4.0.0-rc.0",
+    "@skyux-sdk/e2e": "4.0.0-rc.1",
+    "@skyux-sdk/testing": "4.0.0-rc.3",
+    "@skyux/assets": "4.0.0",
+    "@skyux/config": "4.0.0",
+    "@skyux/core": "4.1.0",
+    "@skyux/http": "4.0.0",
+    "@skyux/i18n": "4.0.0",
+    "@skyux/indicators": "4.0.0",
+    "@skyux/modals": "4.0.0",
+    "@skyux/omnibar-interop": "4.0.0",
+    "@skyux/router": "4.0.0",
+    "@skyux/theme": "4.0.0",
+    "codelyzer": "5.2.2",
+    "rxjs": "6.5.5",
+    "ts-node": "8.3.0",
+    "tslint": "6.1.2",
+    "typescript": "3.8.3",
+    "zone.js": "0.10.3"
   },
   "repository": {
     "type": "git",
