@@ -4,7 +4,7 @@ import {
 
 import {
   SkyClipboardModule
-} from './public';
+} from './public/public_api';
 
 @NgModule({
   exports: [
